Create /etc/timezone and /var/log/www only if missing

+++ init.sh
@@ -1,6 +1,10 @@
-cp /vagrant/vendor/sharkodlak/development/filesystem/etc/timezone /etc/timezone
-chmod 644 /etc/timezone
-chown root:root /etc/timezone
-dpkg-reconfigure --frontend noninteractive tzdata
+if [ ! -e /etc/timezone ]; then
+	cp /vagrant/vendor/sharkodlak/development/filesystem/etc/timezone /etc/timezone
+	chmod 644 /etc/timezone
+	chown root:root /etc/timezone
+	dpkg-reconfigure --frontend noninteractive tzdata
+fi
 
-mkdir --mode=774 /var/log/www
+if [ ! -d /var/log/www ]; then
+	mkdir --mode=774 /var/log/www
+fi
